bgpv1: refactor conversion of family types
Add conversion of API CiliumBGPFamily to agent family type. Also,
refactored gobgp conversions.
